Update: Chat Entity structure

+++ src/main/java/com/docnow/docnow/chat/repository/ChatUserRepository.java
@@ -3,6 +3,6 @@ package com.docnow.docnow.chat.repository;
 import com.docnow.docnow.chat.Entity.ChatUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChatUserRepository extends JpaRepository<ChatUser,Integer> {
+public interface ChatUserRepository extends JpaRepository<ChatUser,Long> {
 
 }
